feat: add inline response chart to AmigoEndpoint
Introduce a new inline chart to display request/response trends in the AmigoEndpoint resource. This chart helps in visualizing the total and failed requests over time, allowing users to quickly identify patterns and issues. The new widget, EndpointResponsesTableChart, aggregates data on a per-day basis for the past month to generate the chart.

- Added `EndpointResponsesTableChart` widget for visualizing endpoint response trends
- Improved readability of responses count in the table by formatting it as a number
- Included the new inline chart in the AmigoEndpointResource page and updated related widgets
- Integrated `lara-zeus/inline-chart` package for rendering inline charts in tables

This enhancement not only improves the user experience by providing a clear visual representation of the endpoint performance but also streamlines the analysis process by offering quick insights at a glance.